functioning visit/patient model views without delete
