+++ project/npda/templates/visits.html
@@ -18,10 +18,10 @@
                     {% if visits %}
                         {% for visit in visits %}
                           <tr class="border-b bg-white text-rcpch_light_blue border-gray-100 hover:bg-r hover:bg-rcpch_light_blue hover:text-white">
-                              <td><a href="{% url 'patient-update' patient.pk %}" class="px-2 flex text-center hover:text-rcpch_pink"><strong>{{ visit.pk }}</strong></a></td>
+                              <td><strong>{{ visit.pk }}</strong></td>
                               <td class="px-6 py-4 text-center">{{ visit.visit_date }}</td>
                               <td>
-                                <a href="{% url 'patient_visits' patient.pk %}" class="text-rcpch_pink px-2 flex justify-center items-center">
+                                <a href="{% url 'visit-update' patient.pk visit.pk %}" class="text-rcpch_pink px-2 flex justify-center items-center">
                                   <strong>Visit Details</strong>
                                   <svg class="h-8 w-8 text-rcpch_pink"  width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">  <path stroke="none" d="M0 0h24v24H0z"/>  <polyline points="9 6 15 12 9 18" /></svg>
                                 </a>
@@ -45,7 +45,7 @@
                     </tfoot>
                   </table>
                 </div>
-                <a href="{% url 'patient-add' %}" class="bg-rcpch_light_blue text-white font-semibold hover:text-white py-2.5 px-3 mt-40 border border-rcpch_light_blue hover:bg-rcpch_strong_blue hover:border-rcpch_strong_blue">Create New Visit</a>
+                <a href="{% url 'visit-create' patient_id=patient.pk %}" class="bg-rcpch_light_blue text-white font-semibold hover:text-white py-2.5 px-3 mt-40 border border-rcpch_light_blue hover:bg-rcpch_strong_blue hover:border-rcpch_strong_blue">Create New Visit</a>
                 <a href="{% url 'patients' %}" class="bg-rcpch_light_blue text-white font-semibold hover:text-white py-2.5 px-3 mt-40 border border-rcpch_light_blue hover:bg-rcpch_strong_blue hover:border-rcpch_strong_blue">Back to Patient List</a>
         </div>
     </div>
